Now we can use the CFramebuffer object more easily and few more other stuff
-Function added to renderer
-addOffscreen()
-getOffscreen()
-Function added to helper
-nameUsed() to check if a name is already used or not
-iterator() to iterate through a collection via string

+++ VulkanEngine/Panel.cpp
@@ -13,5 +13,5 @@ Panel::~Panel()
 
 glm::vec4 Panel::getColor() const
 {
-	return (glm::vec4)m_color/glm::vec4(m_colorRef);
+	return (glm::vec4)m_color/glm::vec4(static_cast<float>(m_colorRef));
 }
